Ensure CLI and GUI config locations are consistent
